requirements.txt
aiohappyeyeballs==2.4.4
aiohttp==3.11.11
aiosignal==1.3.2
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
attrs==24.3.0
bcrypt==4.2.1
certifi==2024.12.14
cffi==1.17.1
click==8.1.8
cryptography==44.0.0
Deprecated==1.2.15
distro==1.9.0
fastapi==0.115.6
frozenlist==1.5.0
greenlet==3.1.1
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.5
jiter==0.8.2
limits==4.0.0
loguru==0.7.3
Mako==1.3.8
MarkupSafe==3.0.2
multidict==6.1.0
mypy-extensions==1.0.0
openai==1.60.2
packaging==24.2
passlib==1.7.4
propcache==0.2.1
psutil==5.9.8
pycparser==2.22
pydantic==2.10.5
pydantic-settings==2.7.1
pydantic_core==2.27.2
PyJWT==2.10.1
python-dotenv==1.0.1
python-multipart==0.0.20
redis==5.2.1
replicate==1.0.4
slowapi==0.1.9
sniffio==1.3.1
sqladmin==0.20.1
SQLAlchemy==2.0.37
starlette==0.41.3
tqdm==4.67.1
typing-inspect==0.9.0
typing_extensions==4.12.2
uvicorn==0.34.0
wrapt==1.17.2
WTForms==3.1.2
yarl==1.18.3
